# docs/swagger.json
{
    "schemes": [
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST API documentation for the Passion Tree learning platform.\nUse the \"Authorize\" button and paste a valid JWT (without the \"Bearer \" prefix) to call protected endpoints.",
        "title": "Passion Tree Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Passion Tree Team",
            "url": "https://passion-tree.org"
        },
        "license": {
            "name": "Proprietary"
        },
        "version": "1.0"
    },
    "host": "api.passion-tree.org",
    "basePath": "/api/v1",
    "paths": {
...
        "/auth/change-password": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Updates the password for the JWT-authenticated user after verifying the old password.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Change password (authenticated)",
                "parameters": [
                    {
                        "description": "Old and new passwords",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passiontree_internal_auth_model.ChangePasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidoc.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Wrong old password or invalid session",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/forgot-password": {
            "post": {
                "description": "Always returns 200 to avoid leaking whether an email exists in the system.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Request password reset email",
                "parameters": [
                    {
                        "description": "Email to send reset code to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passiontree_internal_auth_model.ForgotPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidoc.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/resend-verification": {
            "post": {
                "description": "Sends a fresh verification code to the supplied email if the account exists and is unverified.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Resend the email-verification code",
                "parameters": [
                    {
                        "description": "Email to resend to",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passiontree_internal_auth_model.ResendVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidoc.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/auth/reset-password": {
            "post": {
                "description": "Updates the user's password if the supplied reset code is valid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password using a reset code",
                "parameters": [
                    {
                        "description": "Reset code and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passiontree_internal_auth_model.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidoc.MessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid code or password",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/auth/verify-email": {
            "post": {
                "description": "Confirms a user's email using the one-time code sent at registration and immediately issues access \u0026 refresh tokens.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify email and auto-login",
                "parameters": [
                    {
                        "description": "Verification code",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/passiontree_internal_auth_model.VerifyEmailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/apidoc.TokenPairResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/apidoc.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "apidoc.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "invalid request body"
                },
                "success": {
                    "type": "boolean",
                    "example": false
                }
            }
        },
        "apidoc.MessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Operation completed successfully"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "apidoc.TokenPair": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "eyJhbGciOi..."
                },
                "refresh_token": {
                    "type": "string",
                    "example": "eyJhbGciOi..."
                }
            }
        },
        "apidoc.TokenPairResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/apidoc.TokenPair"
                },
                "message": {
                    "type": "string",
                    "example": "Login successful"
                },
                "success": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
...
        "passiontree_internal_auth_model.ChangePasswordRequest": {
            "type": "object",
            "properties": {
                "new_password": {
                    "type": "string"
                },
                "old_password": {
                    "type": "string"
                }
            }
        },
...
        "passiontree_internal_auth_model.ForgotPasswordRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
...
        "passiontree_internal_auth_model.ResendVerificationRequest": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "passiontree_internal_auth_model.ResetPasswordRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "new_password": {
                    "type": "string"
                }
            }
        },
...
        "passiontree_internal_auth_model.VerifyEmailRequest": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer\" followed by a space and a valid JWT access token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}